+++ packages/file_selector/file_selector_linux/example/lib/get_multiple_directories_page.dart
@@ -21,10 +21,13 @@ class GetMultipleDirectoriesPage extends StatelessWidget {
       // Operation was canceled by the user.
       return;
     }
-    await showDialog<void>(
-      context: context,
-      builder: (BuildContext context) => TextDisplay(directoryPaths.join('\n')),
-    );
+    if (context.mounted) {
+      await showDialog<void>(
+        context: context,
+        builder: (BuildContext context) =>
+            TextDisplay(directoryPaths.join('\n')),
+      );
+    }
   }
 
   @override