+++ packages/camera/camera/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 0.10.3
+## 0.10.4
 
 * Allows camera to be switched while video recording.
+
+## 0.10.3
+
+* Adds back use of Optional type.
 
 ## 0.10.2+1