+++ packages/camera/camera_android/CHANGELOG.md
@@ -1,6 +1,11 @@
-## 0.10.3
+## 0.10.4
 
 * Allows camera to be switched while video recording.
+
+## 0.10.3
+
+* Adds back use of Optional type.
+* Updates minimum Flutter version to 3.0.
 
 ## 0.10.2+3