+++ packages/google_sign_in/google_sign_in/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 5.4.4
+
+* Adds documentation for iOS auth with SERVER_CLIENT_ID
+* Updates minimum Flutter version to 3.0.
+
 ## 5.4.3
 
 * Updates code for stricter lint checks.